+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 - Add `network_id` to `WalletData` so that it is saved with the seed data and surfaced via the export function
+- Add ability to import external wallets into CDP via a BIP-39 mnemonic phrase, as a 1-of-1 wallet
+- Add ability to import WalletData files exported by the NodeJS CDP SDK
+- Deprecate `Wallet.load_seed` method in favor of `Wallet.load_seed_from_file`
+- Deprecate `Wallet.saveSeed()` method in favor of `Wallet.save_seed_to_file`
 
 ### [0.12.1] - 2024-12-10